pyproject.toml
[tool.poetry]
name = "schedule"
version = "0.1.0"
description = ""
authors = ["poldy <poldy200@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
ruff = "^0.3.7"
pytest = "^8.1.1"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 100
select = [
    "E",
    "F",
    "T20",
    "UP",
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".env",
    ".env_legacy",
    "migrations",
    "__init__.py",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.per-file-ignores]
"import/*" = ["T20"]
